requirements.txt
pytest
requests
python-dotenv
black
pytest-xdist
pytest-playwright